requirements.txt
requests
pandas
plotly
scipy
tqdm
python-dotenv
jinja2
anywidget
